+++ lib/src/generator/generator.dart
@@ -1,15 +1,28 @@
+import 'dart:convert';
+import 'dart:io';
+
 import 'package:build/build.dart';
+import 'package:json_locale_generator/src/converter/json_parser.dart';
+import 'package:json_locale_generator/src/errors/invalid_json_format.dart';
 import 'package:json_locale_generator/src/generator/json_class_generator.dart';
 import 'package:json_locale_generator/src/model/resources.dart';
 
 Future<String> generate(BuildStep buildStep, Resources res) async {
-  const ignores =
-      '// ignore_for_file: library_private_types_in_public_api, non_constant_identifier_names';
-
-  final code = <String>[ignores];
-
+  final code = [];
   for (final asset in res.jsonAssets) {
-    code.add(await generateJsonAsset(buildStep, asset));
+    final fileContent = await buildStep.readAsString(asset.id);
+
+    final fileJson = jsonDecode(fileContent) as Map<String, dynamic>;
+    try {
+      final jsonToConvert = jsonCleaner(fileJson, asset.pluralMatcher);
+      code.add(generateParentClass(asset.outputClass, jsonToConvert));
+    } on InvalidJsonKeyFormatException catch (exception) {
+      throw InvalidJsonFormatException(
+        fileContent,
+        File(asset.id.path).absolute.uri,
+        exception.key,
+      );
+    }
   }
   return code.join('\n');
 }