+++ lib/src/utils.dart
@@ -1,4 +1,4 @@
-import 'package:collection/collection.dart';
+
 
 T? safeCast<T>(dynamic obj) {
   if (obj == null) {
@@ -22,22 +22,26 @@ extension StringExt on String {
     final prefix = match?[0];
     if (prefix?.isNotEmpty ?? false) {
       res = res.replaceFirst(prefix!, '');
-      final nonNumericStart = RegExp('[^0-9]*').matchAsPrefix(prefix);
-      final nonNumericPrefixEnd =
-          RegExp(r'[^0-9]+$').allMatches(prefix).lastOrNull;
+      // final nonNumericStart = RegExp('[^0-9]*').matchAsPrefix(prefix);
+      // final nonNumericPrefixEnd =
+      //     RegExp(r'[^0-9]+$').allMatches(prefix).lastOrNull;
       String cleanPrefix = prefix;
       // Remove trailing illegal characters
-      if (nonNumericPrefixEnd?[0]?.isNotEmpty ?? false) {
-        cleanPrefix = cleanPrefix.replaceRange(
-          nonNumericPrefixEnd!.start,
-          nonNumericPrefixEnd.end,
-          '',
-        );
-      }
+      // if (nonNumericPrefixEnd?[0]?.isNotEmpty ?? false) {
+      //   cleanPrefix = cleanPrefix.replaceRange(
+      //     nonNumericPrefixEnd!.start,
+      //     nonNumericPrefixEnd.end,
+      //     '',
+      //   );
+      // }
+      cleanPrefix =
+          cleanPrefix.replaceAllMapped(RegExp('^[^0-9]*'), (match) => '');
+      cleanPrefix =
+          cleanPrefix.replaceAllMapped(RegExp(r'[^0-9]*$'), (match) => '');
       // Check if the key starts with non alphanum characters and if so remove them
-      if (nonNumericStart?[0]?.isNotEmpty ?? false) {
-        cleanPrefix = prefix.replaceFirst(nonNumericStart![0]!, '');
-      }
+      // if (nonNumericStart?[0]?.isNotEmpty ?? false) {
+      //   cleanPrefix = prefix.replaceFirst(nonNumericStart![0]!, '');
+      // }
       cleanPrefix = cleanPrefix.replaceNonAlphaNumChars();
 
       res = '${res}_$cleanPrefix';